updated type handling in assignment expression

+++ testData/expectedResults/InsnNode_i2x.java
@@ -6,7 +6,7 @@ private byte intToByte() {
 int x;
 byte b;
 byte b2;
-x = (int) -45;
+x = -45;
 b = (byte) x;
 b2 = -5;
 return (byte) (b & b2);
