+++ testData/expectedResults/TryCatchBlockNode.java
@@ -13,14 +13,14 @@ try {
 number = java.lang.Integer.valueOf(value).intValue();
 java.lang.System.out.println("no exception");
 } catch (java.lang.NegativeArraySizeException e) {
-number = (int) -6;
+number = -6;
 java.lang.System.out.println("NegativeArraySizeException caught");
 }
 } catch (java.lang.NumberFormatException e) {
 number = -1;
 java.lang.System.out.println("number exception caught");
 } catch (java.lang.IndexOutOfBoundsException e) {
-number = (int) -3;
+number = -3;
 java.lang.System.out.println("index out of bounds exception caught");
 } finally {
 java.lang.System.out.println("called finally");
